Move direct URL scanning helpers to a separate file
